전반적인 클래스 유지보수 및 리팩토링
